refactor: extract SG-008 startup sentinel into startup submodule
Move the SG-008 (ASIL D) process fail-closed startup sentinel —
`StartupContext`, `StartupInvariant`, its `Display` impl, and the pure
`check_startup_invariants` predicate — plus the CERT-003 RTM coverage
tests (`sg_008_cert_tests`) out of the ~3.9K-LOC
`kirra_verifier_service.rs` monolith into a new
`kirra_verifier_service/startup.rs` submodule.

Mechanical, behavior-preserving decomposition consistent with the
existing `#[path = "kirra_verifier_service/X.rs"]` submodule convention
(attestation/fleet/audit/.../operators). The items are re-exported via
`use startup::*;` so `main`'s `StartupContext` build and the
`check_startup_invariants` call site (still immediately before
`TcpListener::bind`) reference them unqualified. The `// Verifies:
SG-008` trace tags, `pub(crate)` visibility, and predicate logic are
moved verbatim — the safety argument and ordering are unchanged.

73 bin tests (incl. 9 SG-008 cert tests) and the traceability gate pass.
